+++ src/components/atoms/icons/CloseIcon.stories.ts
