+++ src/components/atoms/icons/HomeIcon.vue
@@ -10,7 +10,7 @@
 
 <script lang="ts">
 import { defineComponent, computed, PropType, SetupContext } from 'vue';
-import { Size } from './story';
+import { Size } from '../story';
 
 export default defineComponent({
   name: 'HomeIcon',
