+++ src/components/atoms/icons/PauseIcon.stories.ts
