+++ src/components/atoms/icons/RestartIcon.stories.ts
