+++ src/components/atoms/icons/ResumeIcon.stories.ts
