+++ src/components/atoms/story.d.ts
@@ -1,0 +1,2 @@
+// 一旦Sizeだけvalidatorの代用としての型定義
+export type Size = 'small' | 'medium';
