--- src/stories/story.d.ts
@@ -1,1 +1,0 @@
-export type Size = 'small' | 'medium';
